Add confirmation dialog for deletions

+++ frontend/src/app/list/list.component.ts
@@ -1,16 +1,21 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { TodoItem } from '../todo';
 import { TodoService } from '../todo.service';
+import {
+  MatDialog,
+  MatDialogRef,
+  MAT_DIALOG_DATA,
+} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-list',
   templateUrl: './list.component.html',
-  styleUrls: ['./list.component.css']
+  styleUrls: ['./list.component.css'],
 })
 export class ListComponent implements OnInit {
   todoItems: TodoItem[] = [];
 
-  constructor(private todoService: TodoService) { }
+  constructor(private todoService: TodoService, private dialog: MatDialog) {}
 
   ngOnInit(): void {
     this.todoService.getTodoItems().subscribe((data) => {
@@ -19,7 +24,16 @@ export class ListComponent implements OnInit {
   }
 
   deleteItem(item: TodoItem): void {
-    this.todoService.deleteTodoItem(item);
+    const dialogRef = this.dialog.open(ConfirmDeletionDialog, {
+      width: '600px',
+      data: item,
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.todoService.deleteTodoItem(item);
+      }
+    });
   }
 
   updateTodo(item: TodoItem): void {
@@ -27,3 +41,21 @@ export class ListComponent implements OnInit {
   }
 }
 
+@Component({
+  selector: 'app-confirm-deletion-dialog',
+  templateUrl: 'confirm-deletion-dialog.html',
+})
+export class ConfirmDeletionDialog {
+  constructor(
+    public dialogRef: MatDialogRef<ConfirmDeletionDialog>,
+    @Inject(MAT_DIALOG_DATA) public data: TodoItem
+  ) {}
+
+  confirm(): void {
+    this.dialogRef.close(true);
+  }
+
+  cancel(): void {
+    this.dialogRef.close(false);
+  }
+}
